+++ build.gradle.kts
@@ -1,4 +1,3 @@
-import at.skyhanni.sharedvariables.DownloadBackupRepo
 import at.skyhanni.sharedvariables.MinecraftVersion
 import at.skyhanni.sharedvariables.MultiVersionStage
 import at.skyhanni.sharedvariables.ProjectTarget
@@ -9,6 +8,7 @@ import io.gitlab.arturbosch.detekt.DetektCreateBaselineTask
 import net.fabricmc.loom.task.RunGameTask
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+import skyhannibuildsystem.DownloadBackupRepo
 
 plugins {
     idea
